Add an align mode switch to preserve temporary files created by the native aligner. This is useful to look at MAF files or SAM files generated with real datasets. - Fix the way percentages are printed in compact-file-stats. Decimal numbers are prefered over scientific notation for percentages. - Print some statistics about sequence variations recorded in a compact alignment file. - Added some test case files (reads and reference) for sequence variations.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@22449 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/aligners/BWAAligner.java
@@ -250,6 +250,9 @@ public class BWAAligner extends AbstractAligner {
 
     private int seedLength = DEFAULT_SEED_LENGTH;
 
+
+
+
     public BWAAligner() {
         super();
         extensions = new String[]{"rsa", "rpac", "rbwt", "pac", "bwt", "ann", "amb", "sa"};
@@ -505,9 +508,10 @@ public class BWAAligner extends AbstractAligner {
 
         // convert native alignment into compact reads
         final File[] buildResults = processAlignment(referenceFile, readsFile, outputBasename);
-        FileUtils.deleteQuietly(new File(saiBinaryFilename));
-        FileUtils.deleteQuietly(new File(samBinaryFilename));
-
+        if (!keepTemporaryFiles) {
+            FileUtils.deleteQuietly(new File(saiBinaryFilename));
+            FileUtils.deleteQuietly(new File(samBinaryFilename));
+        }
         return buildResults;
     }
 }
